Update draft-poidt-ccamp-actn-poi-pluggable-usecases-gaps.md
An attempt clarify the pluggable landscape and make it less technology dependent. Pluggables can be
p2p, no Line System: This is the case for the majority of pluggables, typically referred to as "grey". Note there is the use case of break-out cables whereby a 4-WL interface can be split up into 4 fibers in a PON structure.
p2p, Line system optional: This is the case for PAM and Coherent modules, whereby PAM is for short distances and typically p2p.
Programmable: This is the case for most coherent pluggables and newer PAM modules, as well as for "grey" optics to distinguish a p2p from a breakout configuration.